requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.2.0
python-decouple==3.8
